# openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080
info:
  title: Messages API
  version: 1.0.0
paths:
  /messages:
    post:
      summary: Send a Message
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: The text you want to send
      responses:
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: integer
                  errors:
                    type: "array"
                    items:
                      type: "string"
        '200':
          description: Successfully sent
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
    get:
      summary: Get Messages
      description: Retrieves a list of messages.
      parameters:
        - name: page
          in: query
          description: Pagination
          required: false
          schema:
            type: "integer"
        - name: status
          in: query
          description: Filter messages by status
          required: false
          schema:
            type: "integer"
            enum:
              - 1 #case SENT;
              - 2 #case READ;
      responses:
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: integer
                  errors:
                    type: "array"
                    items:
                      type: "string"
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  total:
                    type: integer
                  data:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                        uuid:
                          type: "string"
                        text:
                          type: "string"
                        status:
                          type: "string"
                          enum:
                            - 1 #case SENT;
                            - 2 #case READ;
              example:
                total:
                  100
                messages:
                  - text: "Hello, World!"
                    status: "2"
                  - text: "How are you today?"
                    status: "1"
                  - text: "Thanks, I'm fine. How about you?"
                    status: "1"